Refactor DataFormatInfo to contain format size as fraction.

+++ tt_metal/tt-llk/tests/python_tests/helpers/utils.py
@@ -104,7 +104,7 @@ def run_shell_command(
 
 
 def calculate_read_byte_count(format: FormatConfig, array_size: int, sfpu=False) -> int:
-    total_bytes = array_size * format.output_format.size
+    total_bytes = int(array_size * format.output_format.size)
     if format.output_format == DataFormat.Bfp8_b:
         total_bytes += total_bytes // 16
     return total_bytes
